put back test for pushback overflow
git-svn-id: https://svn.r-project.org/R/trunk@59172 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/nmath/pnbeta.c
@@ -62,13 +62,13 @@ pnbeta_raw(double x, double o_x, double a, double b, double ncp)
     j = (int) x0;
     do {
 	j++;
-	temp -= gx;
+	temp -= (double) gx;
 	gx *= x * (a + b + j - 1.) / (a + j);
 	q *= c / j;
 	sumq -= q;
 	ax = temp * q;
 	ans += ax;
-	errbd = (temp - gx) * sumq;
+	errbd = (double)((temp - gx) * sumq);
     }
     while (errbd > errmax && j < itrmax + x0);
